+++ src/elements/pearl-chain/pearl-chain-leg/pearl-chain-leg.scss
@@ -116,9 +116,6 @@
 
       & {
         background-color: var(--sbb-pearl-chain-color-past);
-
-        // --sbb-pearl-chain-leg-status: defined in .ts file
-        width: var(--sbb-pearl-chain-status-position);
       }
     }
   }
