+++ src/elements/pearl-chain/pearl-chain/pearl-chain.ssr.spec.ts
@@ -9,9 +9,13 @@ describe(`sbb-pearl-chain ssr`, () => {
   let root: SbbPearlChainElement;
 
   beforeEach(async () => {
-    root = await ssrHydratedFixture(html`<sbb-pearl-chain></sbb-pearl-chain>`, {
-      modules: ['./pearl-chain.js'],
-    });
+    root = await ssrHydratedFixture(
+      html`<sbb-pearl-chain departure="2022-08-18T04:00" arrival="2022-08-18T05:00">
+      </sbb-pearl-chain>`,
+      {
+        modules: ['./pearl-chain.js'],
+      },
+    );
   });
 
   it('renders', () => {
