+++ src/elements/pearl-chain/pearl-chain/pearl-chain.visual.spec.ts
@@ -95,9 +95,7 @@ describe(`sbb-pearl-chain`, () => {
               now=${(c.now ?? new Date('2024-12-01T12:11:00').valueOf()) / 1000}
               marker="static"
             >
-              ${c.legs.map((l) => {
-                return l;
-              })}</sbb-pearl-chain
+              ${c.legs.map((l) => l)}</sbb-pearl-chain
             >
           `);
         }),
